+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ATLAS Navigator Data Changelog
 
+## [1.1.0]() (2022-06-09)
+
+- Updated to use ATLAS Data 4.0.0, with support for outputting multiple `x-mitre-matrix` objects
+- Outputs STIX 2.1 objects for [ATT&CK Workbench](https://github.com/center-for-threat-informed-defense/attack-workbench-frontend) support
+  + Import the outputted STIX as a new collection in the Workbench
+- Supports ATT&CK Navigator 4.6.4
+
+
 ## [1.0.0]() (2022-03-23)
 
 - Moved ATLAS Navigator scripts and docs into their own repository separate from [ATLAS Data](https://github.com/mitre-atlas/atlas-data)